--- src/main/java/fr/openmc/core/features/milestones/base/MilestoneType.java
@@ -1,6 +1,0 @@
-package fr.openmc.core.features.milestones.base;
-
-public enum MilestoneType {
-	CITY,
-	PLAYER
-}
